Added the DataSeriesExporter class. Reworked the class interface of FileExporter. Introduced the AsyncOperation class.

+++ src/plugins/crystalanalysis/modifier/dxa/InterfaceMesh.cpp
@@ -73,7 +73,7 @@ bool InterfaceMesh::createMesh(FloatType maximumNeighborDistance, const Property
 			if(crystalClusters) {
 				std::array<int,4> clusters;
 				for(int v = 0; v < 4; v++)
-					clusters[v] = crystalClusters->getInt(tessellation().vertexIndex(tessellation().cellVertex(cell, v)));
+					clusters[v] = crystalClusters->getInt64(tessellation().vertexIndex(tessellation().cellVertex(cell, v)));
 				std::sort(std::begin(clusters), std::end(clusters));
 				return (*most_common(std::begin(clusters), std::end(clusters)) + 1);
 			}
